refactor: Criando molecula para o notifications menu
